[SCRUM-60] refactor: SuccessResponse 공통 응답 포맷 정의
