+++ src/main/java/com/projectlyrics/server/domain/auth/external/kakao/KakaoSocialService.java
@@ -6,7 +6,7 @@ import com.projectlyrics.server.domain.auth.external.AuthProvider;
 import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginRequest;
 import com.projectlyrics.server.domain.auth.external.kakao.dto.KakaoAccessTokenResponse;
 import com.projectlyrics.server.domain.auth.external.kakao.dto.KakaoUserInfoResponse;
-import com.projectlyrics.server.global.error_code.ErrorCode;
+import com.projectlyrics.server.global.message.ErrorCode;
 import com.projectlyrics.server.global.exception.BusinessException;
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
