+++ src/main/java/com/projectlyrics/server/domain/user/api/UserController.java
@@ -4,8 +4,10 @@ import com.projectlyrics.server.domain.auth.external.dto.request.UserLoginReques
 import com.projectlyrics.server.domain.common.dto.SuccessResponse;
 import com.projectlyrics.server.domain.user.dto.response.LoginResponse;
 import com.projectlyrics.server.domain.user.usecase.command.UserAuthUseCase;
+import com.projectlyrics.server.global.message.SuccessMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -19,12 +21,14 @@ public class UserController implements UserControllerSwagger {
   private final UserAuthUseCase userAuthUseCase;
 
   @PostMapping("/login/oauth2/kakao")
-  public SuccessResponse<LoginResponse> login(
+  public ResponseEntity<SuccessResponse<LoginResponse>> login(
       @RequestBody UserLoginRequest loginRequest
   ) {
-    return SuccessResponse.of(
-            HttpStatus.CREATED.value(),
+    return ResponseEntity
+        .status(HttpStatus.CREATED)
+        .body(SuccessResponse.of(
+            SuccessMessage.LOGIN_SUCCESS,
             userAuthUseCase.login(loginRequest)
-    );
+        ));
   }
 }
