+++ src/main/java/com/projectlyrics/server/global/exception/BusinessException.java
@@ -1,6 +1,6 @@
 package com.projectlyrics.server.global.exception;
 
-import com.projectlyrics.server.global.error_code.ErrorCode;
+import com.projectlyrics.server.global.message.ErrorCode;
 import lombok.Getter;
 
 @Getter
