+++ src/main/java/com/projectlyrics/server/global/message/SuccessMessage.java
@@ -1,0 +1,14 @@
+package com.projectlyrics.server.global.message;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum SuccessMessage {
+
+  LOGIN_SUCCESS("소셜 로그인이 완료되었습니다."),
+  ;
+
+  private final String message;
+}
